watchmate/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
PyYAML==6.0.1
sqlparse==0.5.0
tzdata==2024.1
